+++ app/login/form.tsx
@@ -1,11 +1,11 @@
-import { HStack } from '@/components/basic/stack';
-import Input from '@/components/basic/input';
-import ErrorLabel from '@/components/basic/errorlabel';
-import { VStack } from '@/components/basic/stack';
-import Radio from '@/components/basic/radio';
-import A from '@/components/basic/a';
-import Button from '@/components/basic/button';
-import Spinner from '@/components/basic/spinner';
+import { HStack } from '@/component/basic/stack';
+import Input from '@/component/basic/input';
+import ErrorLabel from '@/component/basic/errorlabel';
+import { VStack } from '@/component/basic/stack';
+import Radio from '@/component/basic/radio';
+import A from '@/component/basic/a';
+import Button from '@/component/basic/button';
+import Spinner from '@/component/basic/spinner';
 
 interface LoginErrors {
   shake: boolean;
