+++ app/login/reset-pw/inner/form2.tsx
@@ -1,8 +1,8 @@
-import { HStack } from '@/components/basic/stack';
+import { HStack } from '@/component/basic/stack';
 
-import Input from '@/components/basic/input';
-import Button from '@/components/basic/button';
-import Spinner from '@/components/basic/spinner';
+import Input from '@/component/basic/input';
+import Button from '@/component/basic/button';
+import Spinner from '@/component/basic/spinner';
 
 export default function ResetPwForm2({ input, handleInput, handleSubmit, submitting }: FormProps<ResetPwReq>) {
   return (
