+++ app/notices/[index]/page.tsx
@@ -1,4 +1,4 @@
-import { HStack } from '@/components/basic/stack';
+import { HStack } from '@/component/basic/stack';
 import Link from 'next/link';
 import { ELEM_PER_PAGE } from '@/lib/directive';
 import apiGetNotices from '@/lib/api/calls/notice';
