+++ app/register/steps/step1.tsx
@@ -2,17 +2,16 @@
 
 import { useState, useEffect } from 'react';
 
-import Button from '@/components/basic/button';
-import Input from '@/components/basic/input';
-import ErrorLabel from '@/components/basic/errorlabel';
+import Button from '@/component/basic/button';
+import Input from '@/component/basic/input';
+import ErrorLabel from '@/component/basic/errorlabel';
 
-import { HStack, VStack } from '@/components/basic/stack';
+import { HStack, VStack } from '@/component/basic/stack';
+import { RightIcon } from '@/component/icon';
+import Spinner from '@/component/basic/spinner';
 
 import { useAnimationTimeout } from '@/lib/hooks/timeout';
-
-import { RightIcon } from '@/lib/icons';
 import { apiSendEmail } from '@/lib/api/calls/login';
-import Spinner from '@/components/basic/spinner';
 
 const validateEmail = (email: string) => {
   const re = /^[^\s@]+@korea\.ac\.kr$/;
